Use a general component for selecting categories with file
