+++ src/pages/editor/CategoriesWithFiles.tsx
@@ -1,32 +1,15 @@
 import { LoadingButton } from '@mui/lab';
-import { Box, Typography } from '@mui/material';
+import { Typography } from '@mui/material';
 import { useMutation } from '@tanstack/react-query';
 import { useEffect, useState } from 'react';
 import { Helmet } from 'react-helmet-async';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 import { updateCustomerInstitution } from '../../api/customerInstitutionsApi';
+import { CategorySelector } from '../../components/CategorySelector';
 import { setNotification } from '../../redux/notificationSlice';
 import { RootState } from '../../redux/store';
-import {
-  ArtisticType,
-  BookType,
-  ChapterType,
-  DegreeType,
-  ExhibitionContentType,
-  JournalType,
-  MediaType,
-  OtherRegistrationType,
-  PresentationType,
-  PublicationType,
-  ReportType,
-  ResearchDataType,
-} from '../../types/publicationFieldNames';
 import { PublicationInstanceType } from '../../types/registration.types';
-import {
-  RegistrationTypeElement,
-  RegistrationTypesRow,
-} from '../registration/resource_type_tab/components/RegistrationTypesRow';
 
 export const CategoriesWithFiles = () => {
   const dispatch = useDispatch();
@@ -45,15 +28,6 @@ export const CategoriesWithFiles = () => {
     onSuccess: () => dispatch(setNotification({ message: t('feedback.success.update_customer'), variant: 'success' })),
     onError: () => dispatch(setNotification({ message: t('feedback.error.update_customer'), variant: 'error' })),
   });
-
-  const searchValue = '';
-
-  const filterRegistrationTypes = (registrationTypes: RegistrationTypeElement[]) => {
-    const lowerCaseSearchValue = searchValue.toLowerCase();
-    return registrationTypes.filter((registrationType) =>
-      registrationType.text.toLowerCase().includes(lowerCaseSearchValue)
-    );
-  };
 
   const onSelectType = (registrationType: PublicationInstanceType) => {
     if (selectedCategories.includes(registrationType)) {
@@ -74,135 +48,7 @@ export const CategoriesWithFiles = () => {
 
       <Typography sx={{ my: '2rem' }}>{t('editor.categories_with_files_description')}</Typography>
 
-      <Box
-        sx={{
-          display: 'grid',
-          gridTemplateColumns: { xs: '1fr', md: 'auto 1fr' },
-          gap: '1rem',
-          alignItems: 'center',
-        }}>
-        <RegistrationTypesRow
-          mainType={PublicationType.PublicationInJournal}
-          registrationTypes={filterRegistrationTypes(
-            Object.values(JournalType).map((registrationType) => ({
-              value: registrationType,
-              text: t(`registration.publication_types.${registrationType}`),
-              selected: selectedCategories.includes(registrationType),
-            }))
-          )}
-          onChangeType={onSelectType}
-        />
-        <RegistrationTypesRow
-          mainType={PublicationType.Book}
-          registrationTypes={filterRegistrationTypes(
-            Object.values(BookType).map((registrationType) => ({
-              value: registrationType,
-              text: t(`registration.publication_types.${registrationType}`),
-              selected: selectedCategories.includes(registrationType),
-            }))
-          )}
-          onChangeType={onSelectType}
-        />
-        <RegistrationTypesRow
-          mainType={PublicationType.Report}
-          registrationTypes={filterRegistrationTypes(
-            Object.values(ReportType).map((registrationType) => ({
-              value: registrationType,
-              text: t(`registration.publication_types.${registrationType}`),
-              selected: selectedCategories.includes(registrationType),
-            }))
-          )}
-          onChangeType={onSelectType}
-        />
-        <RegistrationTypesRow
-          mainType={PublicationType.Degree}
-          registrationTypes={filterRegistrationTypes(
-            Object.values(DegreeType).map((registrationType) => ({
-              value: registrationType,
-              text: t(`registration.publication_types.${registrationType}`),
-              selected: selectedCategories.includes(registrationType),
-            }))
-          )}
-          onChangeType={onSelectType}
-        />
-        <RegistrationTypesRow
-          mainType={PublicationType.Anthology}
-          registrationTypes={filterRegistrationTypes(
-            Object.values(ChapterType).map((registrationType) => ({
-              value: registrationType,
-              text: t(`registration.publication_types.${registrationType}`),
-              selected: selectedCategories.includes(registrationType),
-            }))
-          )}
-          onChangeType={onSelectType}
-        />
-        <RegistrationTypesRow
-          mainType={PublicationType.Presentation}
-          registrationTypes={filterRegistrationTypes(
-            Object.values(PresentationType).map((registrationType) => ({
-              value: registrationType,
-              text: t(`registration.publication_types.${registrationType}`),
-              selected: selectedCategories.includes(registrationType),
-            }))
-          )}
-          onChangeType={onSelectType}
-        />
-        <RegistrationTypesRow
-          mainType={PublicationType.Artistic}
-          registrationTypes={filterRegistrationTypes(
-            Object.values(ArtisticType).map((registrationType) => ({
-              value: registrationType,
-              text: t(`registration.publication_types.${registrationType}`),
-              selected: selectedCategories.includes(registrationType),
-            }))
-          )}
-          onChangeType={onSelectType}
-        />
-        <RegistrationTypesRow
-          mainType={PublicationType.MediaContribution}
-          registrationTypes={filterRegistrationTypes(
-            Object.values(MediaType).map((registrationType) => ({
-              value: registrationType,
-              text: t(`registration.publication_types.${registrationType}`),
-              selected: selectedCategories.includes(registrationType),
-            }))
-          )}
-          onChangeType={onSelectType}
-        />
-        <RegistrationTypesRow
-          mainType={PublicationType.ResearchData}
-          registrationTypes={filterRegistrationTypes(
-            Object.values(ResearchDataType).map((registrationType) => ({
-              value: registrationType,
-              text: t(`registration.publication_types.${registrationType}`),
-              selected: selectedCategories.includes(registrationType),
-            }))
-          )}
-          onChangeType={onSelectType}
-        />
-        <RegistrationTypesRow
-          mainType={PublicationType.ExhibitionContent}
-          registrationTypes={filterRegistrationTypes(
-            Object.values(ExhibitionContentType).map((registrationType) => ({
-              value: registrationType,
-              text: t(`registration.publication_types.${registrationType}`),
-              selected: selectedCategories.includes(registrationType),
-            }))
-          )}
-          onChangeType={onSelectType}
-        />
-        <RegistrationTypesRow
-          mainType={PublicationType.GeographicalContent}
-          registrationTypes={filterRegistrationTypes(
-            Object.values(OtherRegistrationType).map((registrationType) => ({
-              value: registrationType,
-              text: t(`registration.publication_types.${registrationType}`),
-              selected: selectedCategories.includes(registrationType),
-            }))
-          )}
-          onChangeType={onSelectType}
-        />
-      </Box>
+      <CategorySelector selectedCategories={selectedCategories} onCategoryClick={onSelectType} />
 
       <LoadingButton
         variant="contained"
